Create Json.java
Json.java simplifies the process of creating a JSON object and returning it as a response, similar to popular frameworks like Express.js. You can easily add key-value pairs to the JSON object and retrieve the value associated with a given key. Additionally, this class generates a JSON format response without the need to install any extra dependencies or write excessive code.